Register plugin DI dependencies in indexer entry point so plugin manager can resolve ILogger
